+++ include/zephyr/drivers/misc/flexram/nxp_flexram.h
