+++ samples/boards/nxp/mimxrt1170_evk_cm7/magic_addr/CMakeLists.txt
@@ -5,6 +5,4 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(magic_addr)
 
-zephyr_library_include_directories(${ZEPHYR_BASE}/drivers/memc)
-
 target_sources(app PRIVATE src/main.c)
